pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mcp-visual-design-service"
version = "0.1.0"
description = "MCP Visual Design Service for AI movie generation visual assets"
authors = ["Auto-Generated"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
httpx = "^0.25.0"
python-multipart = "^0.0.9"
requests = "^2.31.0"
Pillow = "^10.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
pytest-httpx = "^0.20.0"
ruff = "^0.1.0"
mypy = "^1.7.0"
ipython = "^8.12.0"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "W", "C", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true